dedicated command to collect the schema definitions
- refresh_schema is in charge to retrieve the schema from AWS.
- refresh_modules now reads a local copy of the schema definitions.

This allow us to:
1. run `refresh_modules` in the CI without any AWS credentials.
2. harmonize the way the module generators work
