+++ amazon_cloud_code_generator/api_specifications/AWS::CloudWatch::MetricStream.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::CloudWatch::MetricStream", "readOnlyProperties": ["arn", "creation_date", "last_update_date", "state"], "description": "Resource Type definition for Metric Stream", "anyOf": [{"required": ["FirehoseArn", "RoleArn", "OutputFormat"]}, {"allOf": [{"required": ["FirehoseArn", "RoleArn", "OutputFormat"]}]}, {"oneOf": [{"required": ["IncludeFilters"]}, {"required": ["ExcludeFilters"]}]}], "createOnlyProperties": ["name"], "primaryIdentifier": ["name"], "required": ["firehose_arn", "role_arn", "output_format"], "sourceUrl": "https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-cloudwatch.git", "handlers": {"read": {"permissions": ["cloudwatch:GetMetricStream"]}, "create": {"permissions": ["cloudwatch:PutMetricStream", "iam:PassRole"]}, "update": {"permissions": ["cloudwatch:PutMetricStream", "iam:PassRole"]}, "list": {"permissions": ["cloudwatch:ListMetricStreams"]}, "delete": {"permissions": ["cloudwatch:DeleteMetricStream"]}}, "writeOnlyProperties": ["tags"], "additionalProperties": false, "definitions": {"MetricStreamStatisticsConfiguration": {"description": "This structure specifies a list of additional statistics to stream, and the metrics to stream those additional statistics for. All metrics that match the combination of metric name and namespace will be streamed with the extended statistics, no matter their dimensions.", "additionalProperties": false, "type": "object", "properties": {"IncludeMetrics": {"maxItems": 100, "uniqueItems": true, "description": "An array that defines the metrics that are to have additional statistics streamed.", "type": "array", "items": {"$ref": "#/definitions/MetricStreamStatisticsMetric"}}, "AdditionalStatistics": {"maxItems": 20, "uniqueItems": true, "description": "The additional statistics to stream for the metrics listed in IncludeMetrics.", "type": "array", "items": {"type": "string"}}}, "required": ["AdditionalStatistics", "IncludeMetrics"]}, "MetricStreamFilter": {"description": "This structure defines the metrics that will be streamed.", "additionalProperties": false, "type": "object", "properties": {"Namespace": {"minLength": 1, "description": "Only metrics with Namespace matching this value will be streamed.", "type": "string", "maxLength": 255}}, "required": ["Namespace"]}, "MetricStreamStatisticsMetric": {"description": "A structure that specifies the metric name and namespace for one metric that is going to have additional statistics included in the stream.", "additionalProperties": false, "type": "object", "properties": {"MetricName": {"minLength": 1, "description": "The name of the metric.", "type": "string", "maxLength": 255}, "Namespace": {"minLength": 1, "description": "The namespace of the metric.", "type": "string", "maxLength": 255}}, "required": ["MetricName", "Namespace"]}, "Tag": {"description": "Metadata that you can assign to a Metric Stream, consisting of a key-value pair.", "additionalProperties": false, "type": "object", "properties": {"Value": {"minLength": 1, "description": "An optional string, which you can use to describe or define the tag.", "type": "string", "maxLength": 256}, "Key": {"minLength": 1, "description": "A unique identifier for the tag.", "type": "string", "maxLength": 128}}, "required": ["Key"]}}, "properties": {"CreationDate": {"format": "date-time", "description": "The date of creation of the metric stream.", "type": "string"}, "StatisticsConfigurations": {"maxItems": 100, "uniqueItems": true, "description": "By default, a metric stream always sends the MAX, MIN, SUM, and SAMPLECOUNT statistics for each metric that is streamed. You can use this parameter to have the metric stream also send additional statistics in the stream. This array can have up to 100 members.", "type": "array", "items": {"$ref": "#/definitions/MetricStreamStatisticsConfiguration"}}, "FirehoseArn": {"minLength": 20, "description": "The ARN of the Kinesis Firehose where to stream the data.", "type": "string", "maxLength": 2048}, "OutputFormat": {"minLength": 1, "description": "The output format of the data streamed to the Kinesis Firehose.", "type": "string", "maxLength": 255}, "ExcludeFilters": {"maxItems": 1000, "uniqueItems": true, "description": "Define which metrics will be not streamed. Metrics matched by multiple instances of MetricStreamFilter are joined with an OR operation by default. If both IncludeFilters and ExcludeFilters are omitted, all metrics in the account will be streamed. IncludeFilters and ExcludeFilters are mutually exclusive. Default to null.", "type": "array", "items": {"$ref": "#/definitions/MetricStreamFilter"}}, "RoleArn": {"minLength": 20, "description": "The ARN of the role that provides access to the Kinesis Firehose.", "type": "string", "maxLength": 2048}, "Name": {"minLength": 1, "description": "Name of the metric stream.", "type": "string", "maxLength": 255}, "IncludeFilters": {"maxItems": 1000, "uniqueItems": true, "description": "Define which metrics will be streamed. Metrics matched by multiple instances of MetricStreamFilter are joined with an OR operation by default. If both IncludeFilters and ExcludeFilters are omitted, all metrics in the account will be streamed. IncludeFilters and ExcludeFilters are mutually exclusive. Default to null.", "type": "array", "items": {"$ref": "#/definitions/MetricStreamFilter"}}, "State": {"minLength": 1, "description": "Displays the state of the Metric Stream.", "type": "string", "maxLength": 255}, "Arn": {"minLength": 20, "description": "Amazon Resource Name of the metric stream.", "type": "string", "maxLength": 2048}, "Tags": {"maxItems": 50, "uniqueItems": true, "description": "A set of tags to assign to the delivery stream.", "type": "array", "items": {"$ref": "#/definitions/Tag"}}, "LastUpdateDate": {"format": "date-time", "description": "The date of the last update of the metric stream.", "type": "string"}}}
