+++ amazon_cloud_code_generator/api_specifications/AWS::EKS::Addon.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::EKS::Addon", "description": "Resource Schema for AWS::EKS::Addon", "sourceUrl": "https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-eks.git", "definitions": {"Tag": {"description": "A key-value pair to associate with a resource.", "type": "object", "additionalProperties": false, "properties": {"Key": {"type": "string", "description": "The key name of the tag. You can specify a value that is 1 to 127 Unicode characters in length and cannot be prefixed with aws:. You can use any of the following characters: the set of Unicode letters, digits, whitespace, _, ., /, =, +, and -. ", "minLength": 1, "maxLength": 127}, "Value": {"type": "string", "description": "The value for the tag. You can specify a value that is 1 to 255 Unicode characters in length and cannot be prefixed with aws:. You can use any of the following characters: the set of Unicode letters, digits, whitespace, _, ., /, =, +, and -. ", "minLength": 1, "maxLength": 255}}, "required": ["Key", "Value"]}}, "properties": {"ClusterName": {"description": "Name of Cluster", "type": "string", "minLength": 1}, "AddonName": {"description": "Name of Addon", "type": "string", "minLength": 1}, "AddonVersion": {"description": "Version of Addon", "type": "string", "minLength": 1}, "ResolveConflicts": {"description": "Resolve parameter value conflicts", "type": "string", "minLength": 1, "enum": ["NONE", "OVERWRITE", "PRESERVE"]}, "ServiceAccountRoleArn": {"description": "IAM role to bind to the add-on's service account", "type": "string", "minLength": 1}, "Arn": {"description": "Amazon Resource Name (ARN) of the add-on", "type": "string"}, "Tags": {"description": "An array of key-value pairs to apply to this resource.", "type": "array", "uniqueItems": true, "insertionOrder": false, "items": {"$ref": "#/definitions/Tag"}}}, "additionalProperties": false, "required": ["cluster_name", "addon_name"], "primaryIdentifier": ["cluster_name", "addon_name"], "createOnlyProperties": ["cluster_name", "addon_name"], "readOnlyProperties": ["arn"], "writeOnlyProperties": ["resolve_conflicts"], "handlers": {"create": {"permissions": ["eks:CreateAddon", "eks:DescribeAddon", "eks:TagResource", "iam:PassRole"]}, "read": {"permissions": ["eks:DescribeAddon"]}, "delete": {"permissions": ["eks:DeleteAddon", "eks:DescribeAddon"]}, "list": {"permissions": ["eks:ListAddons"]}, "update": {"permissions": ["iam:PassRole", "eks:UpdateAddon", "eks:DescribeAddon", "eks:DescribeUpdate", "eks:ListTagsForResource", "eks:TagResource", "eks:UntagResource"]}}}
