+++ amazon_cloud_code_generator/api_specifications/AWS::IAM::ServerCertificate.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::IAM::ServerCertificate", "description": "Resource Type definition for AWS::IAM::ServerCertificate", "additionalProperties": false, "properties": {"CertificateBody": {"minLength": 1, "maxLength": 16384, "pattern": "[\\u0009\\u000A\\u000D\\u0020-\\u00FF]+", "type": "string"}, "CertificateChain": {"minLength": 1, "maxLength": 2097152, "pattern": "[\\u0009\\u000A\\u000D\\u0020-\\u00FF]+", "type": "string"}, "ServerCertificateName": {"minLength": 1, "maxLength": 128, "pattern": "[\\w+=,.@-]+", "type": "string"}, "Path": {"minLength": 1, "maxLength": 512, "pattern": "(\\u002F)|(\\u002F[\\u0021-\\u007F]+\\u002F)", "type": "string"}, "PrivateKey": {"minLength": 1, "maxLength": 16384, "pattern": "[\\u0009\\u000A\\u000D\\u0020-\\u00FF]+", "type": "string"}, "Arn": {"description": "Amazon Resource Name (ARN) of the server certificate", "minLength": 1, "maxLength": 1600, "type": "string"}, "Tags": {"type": "array", "uniqueItems": false, "insertionOrder": false, "items": {"$ref": "#/definitions/Tag"}}}, "definitions": {"Tag": {"description": "A key-value pair to associate with a resource.", "type": "object", "additionalProperties": false, "properties": {"Value": {"description": "The value for the tag. You can specify a value that is 0 to 256 Unicode characters in length and cannot be prefixed with aws:. You can use any of the following characters: the set of Unicode letters, digits, whitespace, _, ., /, =, +, and -.", "type": "string", "minLength": 1, "maxLength": 256}, "Key": {"description": "The key name of the tag. You can specify a value that is 1 to 128 Unicode characters in length and cannot be prefixed with aws:. You can use any of the following characters: the set of Unicode letters, digits, whitespace, _, ., /, =, +, and -.", "type": "string", "minLength": 1, "maxLength": 128}}, "required": ["Value", "Key"]}}, "createOnlyProperties": ["server_certificate_name", "private_key", "certificate_body", "certificate_chain"], "readOnlyProperties": ["arn"], "writeOnlyProperties": ["private_key", "certificate_body", "certificate_chain"], "primaryIdentifier": ["server_certificate_name"], "handlers": {"create": {"permissions": ["iam:UploadServerCertificate", "iam:GetServerCertificate"]}, "read": {"permissions": ["iam:GetServerCertificate"]}, "update": {"permissions": ["iam:TagServerCertificate", "iam:UntagServerCertificate"]}, "delete": {"permissions": ["iam:DeleteServerCertificate"]}, "list": {"permissions": ["iam:ListServerCertificates", "iam:GetServerCertificate"]}}, "tagging": {"taggable": true, "tagOnCreate": true, "tagUpdatable": true, "cloudFormationSystemTags": false}}
