+++ amazon_cloud_code_generator/api_specifications/AWS::KMS::ReplicaKey.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::KMS::ReplicaKey", "description": "The AWS::KMS::ReplicaKey resource specifies a multi-region replica AWS KMS key in AWS Key Management Service (AWS KMS).", "sourceUrl": "https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-kms", "definitions": {"Tag": {"description": "A key-value pair to associate with a resource.", "type": "object", "properties": {"Key": {"type": "string", "description": "The key name of the tag. You can specify a value that is 1 to 128 Unicode characters in length and cannot be prefixed with aws:. You can use any of the following characters: the set of Unicode letters, digits, whitespace, _, ., /, =, +, and -.", "minLength": 1, "maxLength": 128}, "Value": {"type": "string", "description": "The value for the tag. You can specify a value that is 0 to 256 Unicode characters in length and cannot be prefixed with aws:. You can use any of the following characters: the set of Unicode letters, digits, whitespace, _, ., /, =, +, and -.", "minLength": 0, "maxLength": 256}}, "additionalProperties": false, "required": ["Key", "Value"]}}, "properties": {"PrimaryKeyArn": {"description": "Identifies the primary AWS KMS key to create a replica of. Specify the Amazon Resource Name (ARN) of the AWS KMS key. You cannot specify an alias or key ID. For help finding the ARN, see Finding the Key ID and ARN in the AWS Key Management Service Developer Guide.", "type": "string", "minLength": 1, "maxLength": 256}, "Description": {"description": "A description of the AWS KMS key. Use a description that helps you to distinguish this AWS KMS key from others in the account, such as its intended use.", "type": "string", "minLength": 0, "maxLength": 8192}, "Enabled": {"description": "Specifies whether the AWS KMS key is enabled. Disabled AWS KMS keys cannot be used in cryptographic operations.", "type": "boolean"}, "KeyPolicy": {"description": "The key policy that authorizes use of the AWS KMS key. The key policy must observe the following rules.", "type": ["object", "string"]}, "PendingWindowInDays": {"description": "Specifies the number of days in the waiting period before AWS KMS deletes an AWS KMS key that has been removed from a CloudFormation stack. Enter a value between 7 and 30 days. The default value is 30 days.", "type": "integer", "minimum": 7, "maximum": 30}, "Tags": {"description": "An array of key-value pairs to apply to this resource.", "type": "array", "uniqueItems": true, "insertionOrder": false, "items": {"$ref": "#/definitions/Tag"}}, "Arn": {"type": "string"}, "KeyId": {"type": "string"}}, "additionalProperties": false, "required": ["primary_key_arn", "key_policy"], "createOnlyProperties": ["primary_key_arn"], "readOnlyProperties": ["arn", "key_id"], "primaryIdentifier": ["key_id"], "writeOnlyProperties": ["pending_window_in_days"], "taggable": true, "handlers": {"create": {"permissions": ["kms:ReplicateKey", "kms:CreateKey", "kms:DescribeKey", "kms:DisableKey", "kms:TagResource"]}, "read": {"permissions": ["kms:DescribeKey", "kms:GetKeyPolicy", "kms:ListResourceTags"]}, "update": {"permissions": ["kms:DescribeKey", "kms:DisableKey", "kms:EnableKey", "kms:PutKeyPolicy", "kms:TagResource", "kms:UntagResource", "kms:UpdateKeyDescription"]}, "delete": {"permissions": ["kms:DescribeKey", "kms:ScheduleKeyDeletion"]}, "list": {"permissions": ["kms:ListKeys", "kms:DescribeKey"]}}}
