+++ amazon_cloud_code_generator/api_specifications/AWS::RDS::DBProxy.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::RDS::DBProxy", "description": "Resource schema for AWS::RDS::DBProxy", "sourceUrl": "https://github.com/aws-cloudformation/aws-cloudformation-rpdk.git", "definitions": {"AuthFormat": {"type": "object", "properties": {"AuthScheme": {"description": "The type of authentication that the proxy uses for connections from the proxy to the underlying database. ", "type": "string", "enum": ["SECRETS"]}, "Description": {"description": "A user-specified description about the authentication used by a proxy to log in as a specific database user. ", "type": "string"}, "IAMAuth": {"description": "Whether to require or disallow Amazon Web Services Identity and Access Management (IAM) authentication for connections to the proxy. The ENABLED value is valid only for proxies with RDS for Microsoft SQL Server.", "type": "string", "enum": ["DISABLED", "REQUIRED", "ENABLED"]}, "SecretArn": {"description": "The Amazon Resource Name (ARN) representing the secret that the proxy uses to authenticate to the RDS DB instance or Aurora DB cluster. These secrets are stored within Amazon Secrets Manager. ", "type": "string"}, "UserName": {"description": "The name of the database user to which the proxy connects.", "type": "string"}}, "additionalProperties": false}, "TagFormat": {"type": "object", "properties": {"Key": {"type": "string", "pattern": "(\\w|\\d|\\s|\\\\|-|\\.:=+-)*", "maxLength": 128}, "Value": {"type": "string", "pattern": "(\\w|\\d|\\s|\\\\|-|\\.:=+-)*", "maxLength": 128}}, "additionalProperties": false}}, "properties": {"Auth": {"description": "The authorization mechanism that the proxy uses.", "type": "array", "insertionOrder": false, "minItems": 1, "items": {"$ref": "#/definitions/AuthFormat"}}, "DBProxyArn": {"description": "The Amazon Resource Name (ARN) for the proxy.", "type": "string"}, "DBProxyName": {"description": "The identifier for the proxy. This name must be unique for all proxies owned by your AWS account in the specified AWS Region.", "type": "string", "pattern": "[0-z]*", "maxLength": 64}, "DebugLogging": {"description": "Whether the proxy includes detailed information about SQL statements in its logs.", "type": "boolean"}, "Endpoint": {"description": "The endpoint that you can use to connect to the proxy. You include the endpoint value in the connection string for a database client application.", "type": "string"}, "EngineFamily": {"description": "The kinds of databases that the proxy can connect to.", "type": "string", "enum": ["MYSQL", "POSTGRESQL", "SQLSERVER"]}, "IdleClientTimeout": {"description": "The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it.", "type": "integer"}, "RequireTLS": {"description": "A Boolean parameter that specifies whether Transport Layer Security (TLS) encryption is required for connections to the proxy.", "type": "boolean"}, "RoleArn": {"description": "The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in AWS Secrets Manager.", "type": "string"}, "Tags": {"description": "An optional set of key-value pairs to associate arbitrary data of your choosing with the proxy.", "type": "array", "insertionOrder": false, "items": {"$ref": "#/definitions/TagFormat"}}, "VpcId": {"description": "VPC ID to associate with the new DB proxy.", "type": "string"}, "VpcSecurityGroupIds": {"description": "VPC security group IDs to associate with the new proxy.", "type": "array", "insertionOrder": false, "minItems": 1, "items": {"type": "string"}}, "VpcSubnetIds": {"description": "VPC subnet IDs to associate with the new proxy.", "type": "array", "insertionOrder": false, "minItems": 2, "items": {"type": "string"}}}, "additionalProperties": false, "required": ["auth", "db_proxy_name", "engine_family", "role_arn", "vpc_subnet_ids"], "readOnlyProperties": ["db_proxy_arn", "endpoint", "vpc_id"], "createOnlyProperties": ["db_proxy_name", "engine_family", "vpc_subnet_ids"], "primaryIdentifier": ["db_proxy_name"], "handlers": {"create": {"permissions": ["rds:CreateDBProxy", "rds:DescribeDBProxies", "iam:PassRole"]}, "read": {"permissions": ["rds:DescribeDBProxies"]}, "update": {"permissions": ["rds:ModifyDBProxy", "rds:AddTagsToResource", "rds:RemoveTagsFromResource", "iam:PassRole"]}, "delete": {"permissions": ["rds:DescribeDBProxies", "rds:DeleteDBProxy"]}, "list": {"permissions": ["rds:DescribeDBProxies"]}}}
