+++ amazon_cloud_code_generator/api_specifications/AWS::Route53::DNSSEC.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::Route53::DNSSEC", "description": "Resource used to control (enable/disable) DNSSEC in a specific hosted zone.", "properties": {"HostedZoneId": {"description": "The unique string (ID) used to identify a hosted zone.", "type": "string", "pattern": "^[A-Z0-9]{1,32}$"}}, "additionalProperties": false, "createOnlyProperties": ["hosted_zone_id"], "required": ["hosted_zone_id"], "primaryIdentifier": ["hosted_zone_id"], "taggable": false, "handlers": {"create": {"permissions": ["route53:EnableHostedZoneDNSSEC", "kms:DescribeKey", "kms:GetPublicKey", "kms:Sign", "kms:CreateGrant"]}, "read": {"permissions": ["route53:GetDNSSEC"]}, "delete": {"permissions": ["route53:DisableHostedZoneDNSSEC", "kms:DescribeKey", "kms:GetPublicKey", "kms:Sign", "kms:CreateGrant"]}, "list": {"permissions": ["route53:GetDNSSEC", "route53:ListHostedZones"]}}}
