+++ amazon_cloud_code_generator/api_specifications/AWS::S3::MultiRegionAccessPointPolicy.json
@@ -1,0 +1,1 @@
+{"typeName": "AWS::S3::MultiRegionAccessPointPolicy", "description": "The policy to be attached to a Multi Region Access Point", "properties": {"MrapName": {"description": "The name of the Multi Region Access Point to apply policy", "type": "string", "pattern": "^[a-z0-9][-a-z0-9]{1,48}[a-z0-9]$", "minLength": 3, "maxLength": 50}, "Policy": {"description": "Policy document to apply to a Multi Region Access Point", "type": "object"}, "PolicyStatus": {"description": "The Policy Status associated with this Multi Region Access Point", "type": "object", "properties": {"IsPublic": {"type": "string", "description": "Specifies whether the policy is public or not.", "enum": ["true", "false"]}}, "additionalProperties": false, "required": ["IsPublic"]}}, "additionalProperties": false, "required": ["policy", "mrap_name"], "createOnlyProperties": ["mrap_name"], "readOnlyProperties": ["policy_status", "is_public"], "primaryIdentifier": ["mrap_name"], "handlers": {"update": {"permissions": ["s3:PutMultiRegionAccessPointPolicy", "s3:DescribeMultiRegionAccessPointOperation"]}, "read": {"permissions": ["s3:GetMultiRegionAccessPointPolicy", "s3:GetMultiRegionAccessPointPolicyStatus"]}, "list": {"permissions": []}, "delete": {"permissions": ["s3:GetMultiRegionAccessPointPolicy", "s3:GetMultiRegionAccessPoint"]}, "create": {"permissions": ["s3:PutMultiRegionAccessPointPolicy", "s3:DescribeMultiRegionAccessPointOperation"]}}}
