+++ amazon_cloud_code_generator/cmd/refresh_modules.py
@@ -16,19 +16,9 @@ import json
 import traceback
 import copy
 
-BOTO3_IMP_ERR = None
-try:
-    import boto3
-
-    HAS_BOTO3 = True
-except ImportError:
-    BOTO3_IMP_ERR = traceback.format_exc()
-    HAS_BOTO3 = False
-
 from typing import Dict, Iterable, List, Optional, TypedDict
 
 from .resources import RESOURCES
-from .generator import CloudFormationWrapper
 from .generator import generate_documentation
 from .utils import get_module_from_config
 from .utils import camel_to_snake
@@ -231,42 +221,6 @@ def generate_argument_spec(options: Dict) -> str:
     return argument_spec
 
 
-class Schema(TypedDict):
-    """A type for the JSONSchema spec"""
-
-    typeName: str
-    description: str
-    properties: Dict
-    definitions: Optional[Dict]
-    required: Optional[List]
-    primaryIdentifier: List
-    readOnlyProperties: Optional[List]
-    createOnlyProperties: Optional[List]
-    taggable: Optional[bool]
-    handlers: Optional[Dict]
-
-
-def generate_schema(raw_content) -> Dict:
-    json_content = json.loads(raw_content)
-    schema: Dict[str, Schema] = json_content
-
-    for key, value in schema.items():
-        if key != "anyOf":
-            if isinstance(value, list):
-                elems = []
-                for v in value:
-                    if isinstance(v, list):
-                        elems.extend(
-                            [camel_to_snake(p.split("/")[-1].strip()) for p in v]
-                        )
-                    else:
-                        elems.append(camel_to_snake(v.split("/")[-1].strip()))
-
-                schema[key] = elems
-
-    return schema
-
-
 class AnsibleModule:
     template_file = "default_module.j2"
 
@@ -337,15 +291,20 @@ def main():
         default="TODO",
         help="the next major version",
     )
+    parser.add_argument(
+        "--schema-dir",
+        type=pathlib.Path,
+        default=pathlib.Path("amazon_cloud_code_generator/api_specifications"),
+        help="location where to store the collected schemas (default: ./amazon_cloud_code_generator/api_specifications)",
+    )
     args = parser.parse_args()
 
     module_list = []
 
     for type_name in RESOURCES:
-        print("Generating modules")
-        cloudformation = CloudFormationWrapper(boto3.client("cloudformation"))
-        raw_content = cloudformation.generate_docs(type_name)
-        schema = generate_schema(raw_content)
+        print(f"Generating modules {type_name}")
+        schema_file = args.schema_dir / f"{type_name}.json"
+        schema = json.loads(schema_file.read_text())
 
         module = AnsibleModule(schema=schema)
