+++ setup.cfg
@@ -23,5 +23,6 @@ data-files =
 
 [entry_points]
 console_scripts =
+    amazon_cloud_code_generator_refresh_schema = amazon_cloud_code_generator.cmd.refresh_schema:main
     amazon_cloud_code_generator_refresh_modules = amazon_cloud_code_generator.cmd.refresh_modules:main
     amazon_cloud_code_generator_refresh_examples = amazon_cloud_code_generator.cmd.refresh_examples:main
