+++ tests/unit/test_generator.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 import amazon_cloud_code_generator.cmd.generator as g
 import amazon_cloud_code_generator.cmd.refresh_modules as rm
+import amazon_cloud_code_generator.cmd.refresh_schema as rs
 
 
 def resources(filepath):
@@ -49,7 +50,7 @@ def test_Description_normalize():
 
 
 def test_generate_documentation():
-    schema = rm.generate_schema(json.dumps(raw_content))
+    schema = rs.generate_schema(json.dumps(raw_content))
     module = rm.AnsibleModule(schema=schema)
     added_ins = {"module": "1.0.0"}
     documentation = g.generate_documentation(
