+++ tox.ini
@@ -10,6 +10,10 @@ deps = -r{toxinidir}/requirements.txt
 commands =
     pytest -vvv {posargs} {toxinidir}/tests/unit/
 
+[testenv:refresh_schema]
+usedevelop = True
+commands = amazon_cloud_code_generator_refresh_schema --schema-dir {toxinidir}/amazon_cloud_code_generator/api_specifications/
+
 [testenv:black]
 deps =
   black >= 22.0, < 23.0
